post

+++ app/Http/Controllers/TasksController.php
@@ -18,11 +18,18 @@ class TasksController extends Controller
      */
     public function index()
     {
-      $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-      return view('tasks.index', [
-          'tasks' => $tasks,
-      ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        return view('tasks.index', $data);
     }
 
     /**
@@ -52,14 +59,13 @@ class TasksController extends Controller
         'content' => 'required|max:255',
       ]);
 
-      $task = new Task;
-      $task->status = $request->status;
-      $task->content = $request->content;
-
       $user = \Auth::user();
-      $task->user_id = $user->id;
 
-      $task->save();
+      $request->user()->tasks()->create([
+          'content' => $request->content,
+          'status' => $request->status,
+          'user_id' => $user->id,
+      ]);
 
       return redirect('/');
     }
@@ -72,11 +78,14 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-      $task = Task::find($id);
-
-      return view('tasks.show', [
-          'task' => $task,
-      ]);
+      if($this->checkBelongTo($id)){
+          $task = Task::find($id);
+          return view('tasks.show', [
+              'task' => $task,
+          ]);
+      }else{
+          return redirect('/');
+      }
     }
 
     /**
@@ -87,11 +96,14 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-      $task = Task::find($id);
-
-      return view('tasks.edit', [
-          'task' => $task,
-      ]);
+      if($this->checkBelongTo($id)){
+          $task = Task::find($id);
+          return view('tasks.edit', [
+              'task' => $task,
+          ]);
+      }else{
+          return redirect('/');
+      }
     }
 
     /**
@@ -103,15 +115,17 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-      $this->validate($request, [
-        'status' => 'required|max:10',
-        'content' => 'required|max:255',
-      ]);
+      if($this->checkBelongTo($id)){
+          $this->validate($request, [
+            'status' => 'required|max:10',
+            'content' => 'required|max:255',
+          ]);
 
-      $task = Task::find($id);
-      $task->status = $request->status;
-      $task->content = $request->content;
-      $task->save();
+          $task = Task::find($id);
+          $task->status = $request->status;
+          $task->content = $request->content;
+          $task->save();
+      }
 
       return redirect('/');
     }
@@ -124,9 +138,29 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-      $task = Task::find($id);
-      $task->delete();
+      if($this->checkBelongTo($id)){
+          $task = Task::find($id);
+          $task->delete();
+      }
 
       return redirect('/');
     }
+
+    /**
+     * タスクがログインしているユーザーに紐づいているかどうかチェックするプライベートメソッド
+     *
+     * @param  int  $id
+     * @return
+     */
+    private function checkBelongTo($id)
+    {
+        $task = \App\task::find($id);
+
+        if (\Auth::user()->id === $task->user_id) {
+            return true;
+        }else{
+            return false;
+        }
+    }
+
 }
